add file to contain entity-post relations. All methods were before part of wordlift.php
