+++ tests/test-post-entity-relations.php
@@ -110,8 +110,8 @@ class PostEntityRelationsTest extends WP_UnitTestCase {
 	}
         
    
-        function test() {
-
+  /*      function test4W() {
+            
             // Create a post.
             $post_id = wl_create_post( '', 'post', 'post', 'publish', 'post' );
 
@@ -133,20 +133,19 @@ class PostEntityRelationsTest extends WP_UnitTestCase {
             $where_uri = wl_get_entity_uri( $where_id );
 
             // Bind the all 5W to the post (can be passed by both id or uri).
-            wl_5w_set_article_w( $post_id, WL_5W_WHAT, $what_id );
-            wl_5w_set_article_w( $post_id, WL_5W_WHAT, $when_id );  // another entity on the same W
-            wl_5w_set_article_w( $post_id, WL_5W_WHY, $why_uri );   // assign by uri
-            wl_5w_set_article_w( $post_id, WL_5W_WHO, array( $who_id, $what_uri) );    // assign more than one
-            wl_5w_set_article_w( $post_id, WL_5W_WHEN, $when_uri );
-            wl_5w_set_article_w( $post_id, WL_5W_WHERE, $where_id );
+            wl_5w_set_article_w( $post_id, WL_4W_WHAT, $what_id );
+            wl_5w_set_article_w( $post_id, WL_4W_WHAT, $when_id );  // another entity on the same W
+            wl_5w_set_article_w( $post_id, WL_4W_WHO, array( $who_id, $what_uri) );    // assign more than one
+            wl_5w_set_article_w( $post_id, WL_4W_WHEN, $when_uri );   // assign by uri
+            wl_5w_set_article_w( $post_id, WL_4W_WHERE, $where_id );
 
             // Check associations.
             $w5 = wl_5w_get_all_article_Ws( $post_id );
-            $this->assertEquals( array( $what_uri, $when_uri ), $w5[WL_5W_WHAT] );
-            $this->assertEquals( array( $where_uri ), $w5[WL_5W_WHERE] );
-            $this->assertEquals( array( $when_uri ), $w5[WL_5W_WHEN] );
-            $this->assertEquals( array( $what_uri, $when_uri ), wl_5w_get_article_w( $post_id, WL_5W_WHAT ) );
-            $this->assertEquals( array( $where_uri ), wl_5w_get_article_w( $post_id, WL_5W_WHERE ) );
-            $this->assertEquals( array( $who_uri, $what_uri ), wl_5w_get_article_w( $post_id, WL_5W_WHO ) );
-    }
+            $this->assertEquals( array( $what_uri, $when_uri ), $w5[WL_4W_WHAT] );
+            $this->assertEquals( array( $where_uri ), $w5[WL_4W_WHERE] );
+            $this->assertEquals( array( $when_uri ), $w5[WL_4W_WHEN] );
+            $this->assertEquals( array( $what_uri, $when_uri ), wl_5w_get_article_w( $post_id, WL_4W_WHAT ) );
+            $this->assertEquals( array( $where_uri ), wl_5w_get_article_w( $post_id, WL_4W_WHERE ) );
+            $this->assertEquals( array( $who_uri, $what_uri ), wl_5w_get_article_w( $post_id, WL_4W_WHO ) );
+    }*/
 }
